+++ src/pages/blog/tag/[tag]/index.tsx
@@ -1,17 +1,10 @@
 import { BlogScreen } from "@/screens/Blog";
 import { Props as ArticleItemProps } from "@/ui/ArticleItem";
-import { TagType } from "@/ui/Tag";
-import { getBlocks, getDatabase, getPage } from "@/utils/notion";
+import { getDatabase } from "@/utils/notion";
 import { getArticles } from "@/utils/useGetArticles";
 
-export default function TagPage({
-  articles,
-  tags,
-}: {
-  articles: ArticleItemProps[];
-  tags: TagType[];
-}) {
-  return <BlogScreen articles={articles} tags={tags} />;
+export default function TagPage({ articles }: { articles: ArticleItemProps[] }) {
+  return <BlogScreen articles={articles} />;
 }
 
 export async function getStaticPaths() {
@@ -33,31 +26,11 @@ export async function getStaticPaths() {
 }
 
 export async function getStaticProps({ params }: { params: { tag: string } }) {
-  const databaseId = process.env.NOTION_BLOG_DATABASE_ID;
-  const articleDb = await getDatabase(databaseId);
   const articles = await getArticles(params.tag);
-
-  //タグ一覧の取得
-  const tagsList = new Set<string>();
-  const tags: TagType[] = [];
-
-  articleDb.forEach((article: any) => {
-    article.properties.tag.multi_select.forEach((tag: any) => {
-      if (!tagsList.has(tag.name)) {
-        tagsList.add(tag.name);
-        tags.push({
-          id: tag.id,
-          name: tag.name,
-          color: tag.color,
-        });
-      }
-    });
-  });
 
   return {
     props: {
       articles,
-      tags,
     },
   };
 }
