+++ src/screens/Blog/index.module.scss
@@ -1,66 +1,23 @@
-@use '@/styles/variables' as *;
-
-.blogWrapper{
-    display: flex;
-    gap: 48px;
-    justify-content: center;
-    margin: auto;
-    max-width: calc(1280px + 160px);
-    @media(max-width: $mobile){
-        flex-direction: column-reverse;
-        gap: 24px;
-    }
-}
-.side{
-    margin: 64px auto;
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    @media(max-width: calc(360px * 3 + 24px * 2 + 80px * 2)){
-        margin: 40px auto;
-    }
-    @media(max-width: $mobile){
-        // margin: auto;
-    }
-}
-.wrapper_side{
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-
-    .sideTitle{
-        font-size:$text-lg;
-        font-weight: bold;
-    }
-}
-.tags{
-    display: flex;
-    flex-wrap:wrap;
-    max-width: 200px;
-    gap: 4px;
-    @media(max-width: $mobile){
-        max-width:360px;
-    }
-}
+@use "@/styles/variables" as *;
 
 .articlesWrapper {
-    display: grid;
-    max-width: calc(360px * 3 + 24px);
-    grid-template-columns: repeat(3, 1fr);
-    justify-content: center;
-    gap: 32px 24px;
-    margin: 64px auto;
+  display: grid;
+  max-width: calc(360px * 3 + 24px);
+  grid-template-columns: repeat(3, 1fr);
+  justify-content: center;
+  gap: 32px 24px;
+  margin: 64px auto;
 
-    @media (max-width: calc(360px * 3 + 24px * 2 + 80px * 2)) {
-        grid-template-columns: repeat(2, 1fr);
-        max-width: calc(360px * 2 + 24px);
-        margin: 40px auto;
-    }
+  @media (max-width: calc(360px * 3 + 24px * 2 + 80px * 2)) {
+    grid-template-columns: repeat(2, 1fr);
+    max-width: calc(360px * 2 + 24px);
+    margin: 40px auto;
+  }
 
-    @media (max-width: calc(360px * 2 + 24px + 32px * 2)) {
-        grid-template-columns: 1fr;
-        gap: 16px;
-        max-width: 360px;
-        margin: 32px auto;
-    }
+  @media (max-width: calc(360px * 2 + 24px + 32px * 2)) {
+    grid-template-columns: 1fr;
+    gap: 16px;
+    max-width: 360px;
+    margin: 32px auto;
+  }
 }
